backend/requirements.txt
fastapi
uvicorn
numpy
Pillow
torch
opencv-python
insightface
faiss-cpu
firebase-admin
torchvision
tqdm
matplotlib
